requirements.txt
flask
qrcode[pil]
python-barcode
pillow
pyzbar
opencv-python
reportlab
